Dataset/Scrapper/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
sqlalchemy=1.3.16=py37he774522_0
sqlite=3.31.1=he774522_0
psaw==0.0.12
Dataset==1.3.1
Pillow==7.1.2
opencv=3.4.2=py37h40b0b35_0
requests=2.23.0=pypi_0
